+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Utils.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.util;
+package org.firstinspires.ftc.teamcode;
 
 public final class Utils
 {
